chore: move nodes into sub-package
