+++ plugin-core/src/test/java/appland/toolwindow/runtimeAnalysis/RuntimeAnalysisModelTest.java
@@ -4,6 +4,10 @@ import appland.AppMapBaseTest;
 import appland.problemsView.FindingsManager;
 import appland.problemsView.TestFindingsManager;
 import appland.settings.AppMapApplicationSettingsService;
+import appland.toolwindow.runtimeAnalysis.nodes.FindingsTableNode;
+import appland.toolwindow.runtimeAnalysis.nodes.ImpactDomainNode;
+import appland.toolwindow.runtimeAnalysis.nodes.Node;
+import appland.toolwindow.runtimeAnalysis.nodes.RootNode;
 import com.intellij.openapi.application.WriteAction;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
